add transfer method, need cleanup more :(

+++ db/migration/000003_add_wallet.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE "wallets" (
     "id" bigserial PRIMARY KEY,
     "user_id" bigint NOT NULL,
-    "balance" bigint NOT NULL DEFAULT 0,
+    "balance" float(32) NOT NULL DEFAULT 0,
     "currency" varchar NOT NULL,
     "created_at" timestamptz NOT NULL DEFAULT (now())
 );
