+++ db/sqlc/transfer.sql.go
@@ -20,9 +20,9 @@ INSERT INTO transfers (
 `
 
 type CreateTransferParams struct {
-	FromWalletID int64 `json:"from_wallet_id"`
-	ToWalletID   int64 `json:"to_wallet_id"`
-	Amount       int64 `json:"amount"`
+	FromWalletID int64   `json:"from_wallet_id"`
+	ToWalletID   int64   `json:"to_wallet_id"`
+	Amount       float64 `json:"amount"`
 }
 
 func (q *Queries) CreateTransfer(ctx context.Context, arg CreateTransferParams) (Transfer, error) {
